+++ app/src/main/java/com/github/melq/howmanydays/data/repository/DayInfoRepository.kt
@@ -2,7 +2,16 @@ package com.github.melq.howmanydays.data.repository
 
 import com.github.melq.howmanydays.data.dao.DayInfoDao
 import com.github.melq.howmanydays.data.entity.DayInfo
-import com.github.melq.howmanydays.data.repository.interfaces.IDayInfoRepository
+
+interface IDayInfoRepository {
+    suspend fun getAllDayInfos(): List<DayInfo>
+
+    suspend fun insertDayInfo(dayInfo: DayInfo): Long
+
+    suspend fun updateDayInfo(dayInfo: DayInfo)
+
+    suspend fun deleteDayInfo(dayInfo: DayInfo)
+}
 
 class DayInfoRepository(private val dayInfoDao: DayInfoDao) : IDayInfoRepository {
     override suspend fun getAllDayInfos(): List<DayInfo> {
@@ -14,12 +23,7 @@ class DayInfoRepository(private val dayInfoDao: DayInfoDao) : IDayInfoRepository
     }
 
     override suspend fun updateDayInfo(dayInfo: DayInfo) {
-        dayInfoDao.updateDayInfo(
-            dayInfo.id,
-            dayInfo.title,
-            dayInfo.date,
-            dayInfo.displayMode.name
-        )
+        dayInfoDao.updateDayInfo(dayInfo.id, dayInfo.title, dayInfo.date, dayInfo.displayMode.name)
     }
 
     override suspend fun deleteDayInfo(dayInfo: DayInfo) {
